CarType Policy add & Super Admin MIddleware added!

+++ app/Http/Controllers/Admin/CarTypeController.php
@@ -2,8 +2,10 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Models\CarType;
+use Gate;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use function Laravel\Prompts\confirm;
 
 class CarTypeController extends Controller
 {
@@ -12,6 +14,7 @@ class CarTypeController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', CarType::class);
         $car_types = CarType::query()
             ->filter(request(['search']))
             ->latest()
@@ -25,6 +28,7 @@ class CarTypeController extends Controller
      */
     public function create(Request $request)
     {
+        Gate::authorize('create', CarType::class);
         return view('Admin.car-types.create');
     }
 
@@ -55,6 +59,7 @@ class CarTypeController extends Controller
      */
     public function edit(CarType $carType)
     {
+        Gate::authorize('update', $carType);
         return view('admin.car-types.edit', compact('carType'));
     }
 
@@ -63,6 +68,7 @@ class CarTypeController extends Controller
      */
     public function update(Request $request, CarType $carType)
     {
+
         $fields = $request->validate([
             'name' => 'required|string|max:255|unique:car_types,name,' . $carType->id,
         ]);
@@ -78,6 +84,8 @@ class CarTypeController extends Controller
      */
     public function destroy(CarType $carType)
     {
+        Gate::authorize('delete', $carType);
+
         $carType->delete();
         return back()->with('success', "$carType->name delete Success!");
     }
